+++ Blaster-Master/Camera.cpp
@@ -38,17 +38,17 @@ void Camera::UpdateCamera()
 	}
 	float player_x, player_y;
 	CPlayer::GetInstance()->GetPosition(player_x, player_y);
-	float screen_width = CGame::GetInstance()->GetScreenWidth();
-	float screen_height = CGame::GetInstance()->GetScreenHeight();
+	float screen_width = (float) CGame::GetInstance()->GetScreenWidth();
+	float screen_height = (float) CGame::GetInstance()->GetScreenHeight();
 
 	float max_height = 60;
 	float max_width = 30;
 
 	RECT bouncing;
-	bouncing.left = cam_x + screen_width / 2 - max_width;
-	bouncing.right = cam_x + screen_width / 2 + max_width;
-	bouncing.top = cam_y - screen_height / 2 + max_height;
-	bouncing.bottom = cam_y - screen_height / 2 - max_height;
+	bouncing.left = (long) (cam_x + screen_width / 2 - max_width);
+	bouncing.right = (long) (cam_x + screen_width / 2 + max_width);
+	bouncing.top = (long) (cam_y - screen_height / 2 + max_height);
+	bouncing.bottom = (long) (cam_y - screen_height / 2 - max_height);
 
 	RECT bouncingMap;
 	bouncingMap.left = 0;
@@ -59,7 +59,7 @@ void Camera::UpdateCamera()
 	if (player_x < bouncing.left) // move left
 	{
 		cam_x += player_x - bouncing.left;
-		if (cam_x < bouncingMap.left) cam_x = bouncingMap.left;
+		if (cam_x < bouncingMap.left) cam_x = (float) bouncingMap.left;
 	}
 
 	if (player_x > bouncing.right)
@@ -74,7 +74,7 @@ void Camera::UpdateCamera()
 	if (player_y > bouncing.top)
 	{
 		cam_y += player_y - bouncing.top;
-		if (cam_y > bouncingMap.top) cam_y = bouncingMap.top;
+		if (cam_y > bouncingMap.top) cam_y = (float) bouncingMap.top;
 	}
 
 	if (player_y < bouncing.bottom)
